+++ lib/generators/cms/widget/maps/google_maps/google_maps_generator.rb
@@ -14,21 +14,18 @@ module Cms
               Model::ApiGenerator.new(behavior: behavior) do |model|
                 model.name = obj_class_name
                 model.title = 'Widget: GoogleMaps'
+                model.migration_path = "#{widget_path}/migrate"
+                model.model_path = model_path
                 model.attributes = [
                   {
                     name: address_attribute_name,
                     type: :string,
                     title: 'Address',
                   },
-                  {
-                    name: sort_key_attribute_name,
-                    type: :string,
-                    title: 'Sort key',
-                  }
                 ]
               end
 
-              turn_model_into_widget(obj_class_name)
+              turn_model_into_widget(obj_class_name, model_path)
             rescue Cms::Generators::DuplicateResourceError
             end
           end
@@ -36,7 +33,7 @@ module Cms
           def copy_app_directory
             directory('app', force: true)
 
-            template('thumbnail.html.haml', 'app/widgets/google_maps_widget/thumbnail.html.haml')
+            template('thumbnail.html.haml', "#{widget_path}/thumbnail.html.haml")
           end
 
           def notice
@@ -47,16 +44,20 @@ module Cms
 
           private
 
+          def widget_path
+            'app/widgets/google_maps_widget'
+          end
+
+          def model_path
+            'app/models'
+          end
+
           def obj_class_name
             'GoogleMapsWidget'
           end
 
           def address_attribute_name
             'address'
-          end
-
-          def sort_key_attribute_name
-            'sort_key'
           end
         end
       end