+++ lib/generators/cms/widget/person/person_generator.rb
@@ -11,12 +11,9 @@ module Cms
             Model::ApiGenerator.new(behavior: behavior) do |model|
               model.name = obj_class_name
               model.title = 'Widget: Person'
+              model.migration_path = "#{widget_path}/migrate"
+              model.model_path = model_path
               model.attributes = [
-                {
-                  name: sort_key_attribute_name,
-                  type: :string,
-                  title: 'Sort key',
-                },
                 {
                   name: person_attribute_name,
                   type: :string,
@@ -25,7 +22,7 @@ module Cms
               ]
             end
 
-            turn_model_into_widget(obj_class_name)
+            turn_model_into_widget(obj_class_name, model_path)
           rescue Cms::Generators::DuplicateResourceError
           end
         end
@@ -33,7 +30,7 @@ module Cms
         def copy_app_directory
           directory('app', force: true)
 
-          template('thumbnail.html.haml', 'app/widgets/person_widget/thumbnail.html.haml')
+          template('thumbnail.html.haml', "#{widget_path}/thumbnail.html.haml")
         end
 
         def notice
@@ -44,12 +41,16 @@ module Cms
 
         private
 
-        def obj_class_name
-          'PersonWidget'
+        def widget_path
+          'app/widgets/person_widget'
         end
 
-        def sort_key_attribute_name
-          'sort_key'
+        def model_path
+          'app/models'
+        end
+
+        def obj_class_name
+          'PersonWidget'
         end
 
         def person_attribute_name